+++ changelog.md
@@ -4,7 +4,16 @@ Changelog
 0.5 -> 0.6
 ----
 
+## Response
+
+* Added `response.isCompleted()`
+
+## Parser
+
 * `flatmap()` renamed to `flatMap()`
+
+## Flow Bundle
+
 * `F.lazy(parser, params, self )` accepts a third `this` parameter
 * `F.sequence()` is deprecated
 
@@ -14,6 +23,8 @@ Added  to the FlowBundle
 * `F.startsWith(value)`
 * `F.moveUntil(string|stopParser)`
 * `F.dropTo(string|stopParser)`
+
+